consolidate tests with other signature tests
